# back-end/laravel/public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employment Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/employers": {
            "get": {
                "tags": [
                    "EmployerController"
                ],
                "summary": "Obter dados de todos os empregadores",
                "description": "end-point para obter os dados de todos os empregadores",
                "operationId": "f07fd5f25bb797deb33a5fe0b9d9bccd",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagina desejada da paginacao",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna uma lista com objetos contendo os dados dos empregadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": 1
                                            },
                                            "cnpj": {
                                                "type": "string",
                                                "example": "11222333000181"
                                            },
                                            "bussiness_name": {
                                                "type": "string",
                                                "example": "Empresa Tal"
                                            },
                                            "fantasy_name": {
                                                "type": "string",
                                                "example": "Tal Empresa"
                                            },
                                            "user_id": {
                                                "type": "number",
                                                "example": "8"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/employers/{id}": {
            "get": {
                "tags": [
                    "EmployerController"
                ],
                "summary": "Obter dados de um empregador especifico com base no id",
                "description": "end-point para obter os dados referentes a um determinado empregador",
                "operationId": "3d3971adc3dfcb59d32b0cb830847654",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do empregador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna os dados do empregador",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "cnpj": {
                                            "type": "string",
                                            "example": "11222333000181"
                                        },
                                        "bussiness_name": {
                                            "type": "string",
                                            "example": "Empresa Tal"
                                        },
                                        "fantasy_name": {
                                            "type": "string",
                                            "example": "Tal Empresa"
                                        },
                                        "user_id": {
                                            "type": "number",
                                            "example": "8"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "EmployerController"
                ],
                "summary": "Deletar conta do empregador",
                "description": "end-point utilizado para deletar a propria conta empregador (Apenas a conta proprietaria do token de autorizacao pode ser deletada).",
                "operationId": "8a844b978884976ed455cfbe2b070f78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do empregador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbiden (Ao tentar apagar uma conta que nao seja a propria)"
                    },
                    "500": {
                        "description": "Interna Server Error"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        },
        "/api/auth/employer/register": {
            "post": {
                "tags": [
                    "AuthController"
                ],
                "summary": "Registrar um novo empregador",
                "description": "end-point utilizado para registrar um novo empregador",
                "operationId": "7e2c0bfe50086b7ff22ca71b785428b5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "cnpj": {
                                        "type": "string",
                                        "example": "11222333000188"
                                    },
                                    "bussiness_name": {
                                        "type": "string",
                                        "example": "Empresa Tal"
                                    },
                                    "fansasy_name": {
                                        "type": "string",
                                        "example": "Tal Empresa"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Tal Empresa"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "teste_usuario@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna dados do usuario e dados do empregador"
                    },
                    "500": {
                        "description": "Interna Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Baerer token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}